more normmmm

+++ map.c
@@ -6,7 +6,7 @@
 /*   By: jsaintho <jsaintho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 12:17:45 by jsaintho          #+#    #+#             */
-/*   Updated: 2025/01/24 16:26:32 by jsaintho         ###   ########.fr       */
+/*   Updated: 2025/01/24 18:08:39 by jsaintho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,18 +16,18 @@ int	position_to_map_tiles(int px, int py, t_cub3d *f)
 {
 	if (px > WIDTH || py > HEIGHT || px < 0 || py < 0)
 		return (1);
-	if ((int)(px / (WIDTH / (f->map)->width)) >= (f->map)->width
-		|| (int)((py / (HEIGHT / (f->map)->height))) >= (f->map)->height)
+	if ((px / (WIDTH / (f->map)->width)) >= (f->map)->width
+		|| ((py / (HEIGHT / (f->map)->height))) >= (f->map)->height)
 		return (1);
-	if (ft_char_in_set((f->map)->map_matrix[(int)(
-				(py / (HEIGHT / (f->map)->height))
-			)][(int)(
-		(px / (WIDTH / (f->map)->width))
-	)], "1P"))
+	if (ft_char_in_set((f->map)->map_matrix[(
+					(py / (HEIGHT / (f->map)->height))
+				)][(
+				(px / (WIDTH / (f->map)->width))
+			)], "1P"))
 	{
-		f->last_wall = (f->map)->map_matrix[(int)(
+		f->last_wall = (f->map)->map_matrix[(
 				(py / (HEIGHT / (f->map)->height))
-				)][(int)(
+				)][(
 				(px / (WIDTH / (f->map)->width))
 				)];
 		return (1);
